# controllers/startupsController.js
let startups = [
    {
    id:1, 
    name: "GREENCYBER", 
    description: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed nisi metus, faucibus ac lacus sit amet, ullamcorper gravida nunc. Nam sit amet ipsum gravida, sagittis risus vel, hendrerit tellus. Nunc diam mi, posuere vitae ante vitae, gravida eleifend magna. Aenean ultrices velit libero, vitae consequat leo bibendum non. Duis eget lacus mi. Quisque vitae tristique metus. Suspendisse non diam risus. Phasellus dapibus vestibulum malesuada. Aenean venenatis urna et massa ullamcorper porttitor. Duis in diam non turpis pharetra egestas non vitae dolor. Phasellus dui ipsum, tempor a egestas nec, scelerisque a velit. Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas.",
    email: 'GREENCYBER@gmail.com', 
    interestRate:"18%", 
    term:"18 months", 
    type:"Specialized", 
    motto:"Think Different",
    location:"Tirana",
    img: "https://i.pinimg.com/736x/36/a9/ee/36a9eed7d1432d00468cb3a4a99cb38c.jpg"
    },
    {
    id:2, 
    name: "OLDLUX", 
    email: 'OLDLUX@gmail.com',
    description: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed nisi metus, faucibus ac lacus sit amet, ullamcorper gravida nunc. Nam sit amet ipsum gravida, sagittis risus vel, hendrerit tellus. Nunc diam mi, posuere vitae ante vitae, gravida eleifend magna. Aenean ultrices velit libero, vitae consequat leo bibendum non. Duis eget lacus mi. Quisque vitae tristique metus. Suspendisse non diam risus. Phasellus dapibus vestibulum malesuada. Aenean venenatis urna et massa ullamcorper porttitor. Duis in diam non turpis pharetra egestas non vitae dolor. Phasellus dui ipsum, tempor a egestas nec, scelerisque a velit. Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas." ,
    interestRate:"12%", 
    term:"12 months", 
    motto:"A Diamond is Forever",
    type:"Specialized", 
    location:"Sofia",
    img: "https://i.pinimg.com/originals/e6/f9/aa/e6f9aaf18d844334eecde0ed0fe70b76.jpg"
    },
    {
    id:3, 
    name: "WIREDGREEN", 
    email: 'WIREDGREEN@gmail.com', 
    description: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed nisi metus, faucibus ac lacus sit amet, ullamcorper gravida nunc. Nam sit amet ipsum gravida, sagittis risus vel, hendrerit tellus. Nunc diam mi, posuere vitae ante vitae, gravida eleifend magna. Aenean ultrices velit libero, vitae consequat leo bibendum non. Duis eget lacus mi. Quisque vitae tristique metus. Suspendisse non diam risus. Phasellus dapibus vestibulum malesuada. Aenean venenatis urna et massa ullamcorper porttitor. Duis in diam non turpis pharetra egestas non vitae dolor. Phasellus dui ipsum, tempor a egestas nec, scelerisque a velit. Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas." ,
    interestRate:"8%", 
    term:"12 months", 
    motto:"The Breakfast of Champions" ,
    type:"Specialized", 
    location:"Tirana",
    img:"https://pbs.twimg.com/media/EkrdztFWAAAl0EZ.jpg"
    },
    {
    id:4, 
    name: "WISERED", 
    email: 'WISERED@gmail.com', 
    description: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed nisi metus, faucibus ac lacus sit amet, ullamcorper gravida nunc. Nam sit amet ipsum gravida, sagittis risus vel, hendrerit tellus. Nunc diam mi, posuere vitae ante vitae, gravida eleifend magna. Aenean ultrices velit libero, vitae consequat leo bibendum non. Duis eget lacus mi. Quisque vitae tristique metus. Suspendisse non diam risus. Phasellus dapibus vestibulum malesuada. Aenean venenatis urna et massa ullamcorper porttitor. Duis in diam non turpis pharetra egestas non vitae dolor. Phasellus dui ipsum, tempor a egestas nec, scelerisque a velit. Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas." ,
    interestRate:"13%", 
    term:"16 months", 
    motto:"We go the extra mile",
    type:"Specialized", 
    location:"Prague",
    img:"https://www.wiseredllc.com/wp-content/uploads/2020/06/register2.jpg"
    },
    {
    id:5, 
    name: "REDGENICS", 
    email: 'REDGENICS@gmail.com', 
    interestRate:"25%", 
    description: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed nisi metus, faucibus ac lacus sit amet, ullamcorper gravida nunc. Nam sit amet ipsum gravida, sagittis risus vel, hendrerit tellus. Nunc diam mi, posuere vitae ante vitae, gravida eleifend magna. Aenean ultrices velit libero, vitae consequat leo bibendum non. Duis eget lacus mi. Quisque vitae tristique metus. Suspendisse non diam risus. Phasellus dapibus vestibulum malesuada. Aenean venenatis urna et massa ullamcorper porttitor. Duis in diam non turpis pharetra egestas non vitae dolor. Phasellus dui ipsum, tempor a egestas nec, scelerisque a velit. Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas." ,
    term:"36 months", 
    motto:"Belong anywhere",
    type:"Specialized", 
    location:"Moscow",
    img:"https://bucket.credgenics.com/staging/assets/img/startegy.jpg"
    },
    {
    id:6, 
    name: "DOMAINGREEN", 
    email: 'DOMAINGREEN@gmail.com', 
    interestRate:"15%",
    description: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed nisi metus, faucibus ac lacus sit amet, ullamcorper gravida nunc. Nam sit amet ipsum gravida, sagittis risus vel, hendrerit tellus. Nunc diam mi, posuere vitae ante vitae, gravida eleifend magna. Aenean ultrices velit libero, vitae consequat leo bibendum non. Duis eget lacus mi. Quisque vitae tristique metus. Suspendisse non diam risus. Phasellus dapibus vestibulum malesuada. Aenean venenatis urna et massa ullamcorper porttitor. Duis in diam non turpis pharetra egestas non vitae dolor. Phasellus dui ipsum, tempor a egestas nec, scelerisque a velit. Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas.",
    term:"18 months", 
    motto:"Go beyond", //MAX 30 CHARACTERS
    type:"Specialized", 
    location:"London",
    img:"https://progressivetimes.files.wordpress.com/2011/04/green-internet.jpg"
}];


const startupsController = {
    async createStartups(req, res) {
        const {id, name, email} = req.body;
        if(name && email){
            if(startups)
                startups.push({id:id, name: name, email: email});
            res.status(200).send('Request Complete');
        }
        else
            res.status(400).send('All fields are required.');
    },
    
    async deleteStartup(req, res) {
        let toKeepStartups = startups.filter(startup => startup.id != req.params.id);
        startups=toKeepStartups;
        res.json(startups);
    },
    
    async getAllStartups(req, res) {
        res.json({ok: true, startupList: startups});
    },
    
    async getStartup(req, res) {
        const startupToGet = req.params;
        const requestedStartup = startups.filter(startup => startup.id == startupToGet.id)
        res.json(requestedStartup);
    },
    
    async updateStartup(req, res) {
        let arr = startups.filter(startup => {
            const newU = req.body;
            if(startup.id == req.params.id) {
                startup = {...startup, ...newU}
                res.status(200).send(startup);
            }

        });
        if(arr.length == 0) {
            res.status(403).send('No startup with such ID');
        }
    },
}

// async function asyncHandler(fn) {
//     // Because express can't deal with async middleware
//     return (req, res, next) => {
//         fn(req, res, next).catch(next);
//     };
// }
//module.exports.function = asyncHandler(function);

module.exports = startupsController;